+++ src/FastAccelStepper.h
@@ -83,6 +83,15 @@ class FastAccelStepperEngine {
   // the return value of this call.
   FastAccelStepper* stepperConnectToPin(uint8_t step_pin);
 
+  // For e.g. esp32, there are two types of driver.
+  // One using mcpwm and pcnt module. And another using rmt module.
+  // This call allows to select the respective driver
+#if defined(SUPPORT_SELECT_DRIVER_TYPE)
+#define DRIVER_MCPWM_PCNT 0
+#define DRIVER_RMT        1
+  FastAccelStepper* stepperConnectToPin(uint8_t step_pin, uint8_t driver_type);
+#endif
+
   // Comments to valid pins:
   //
   // clang-format off