+++ CHANGELOG.md
@@ -5,6 +5,9 @@ TODO:
 - rename PoorManFloat to e.g. Log2Representation
 - rename RampConstAcceleration to e.g. RampControl
 
+pre-0.29.2:
+- add `stepperConnectToPin(pin, driver_type)` to allow the module selection
+
 0.29.1:
 - improve rounding for log2/pow2 conversion
 - avr: reduce max speed for single stepper use down to 50kSteps/s from 70kSteps/s